// src/app/shared/models/customer-info.ts
export interface ICustomerInfo {
  userName: string;
  name: string;
  idNumber: string;
  phoneNumber: string;
  emailAddress: string;
  registrationDate: string;
  accountBalance: string;
  accountType: string;
  bankName: string;
  status: string;
  id?: string;
  password?: string;
  confirmPassword?: string;

}